Move all shared tests to androidTest and add documentation to README
