+++ test/src/androidTest/kotlin/io/realm/shared/LinkTests.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.realm
+package io.realm.shared
 
 import io.realm.util.PlatformUtils
 import test.link.Child
