+++ test/src/androidTest/kotlin/io/realm/shared/SampleTests.kt
@@ -14,8 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package io.realm
+package io.realm.shared
 
 import io.realm.util.PlatformUtils
 import test.Sample
