--- test/src/macosTest/kotlin/io/realm/ImportTests.kt
@@ -1,1 +1,0 @@
-../../../../androidTest/kotlin/io/realm/ImportTests.kt
