--- test/src/macosTest/kotlin/io/realm/NotificationTests.kt
@@ -1,1 +1,0 @@
-../../../../androidTest/kotlin/io/realm/NotificationTests.kt
