--- test/src/macosTest/kotlin/io/realm/NullabilityTests.kt
@@ -1,1 +1,0 @@
-../../../../androidTest/kotlin/io/realm/NullabilityTests.kt
