--- test/src/macosTest/kotlin/io/realm/PrimaryKeyTests.kt
@@ -1,1 +1,0 @@
-../../../../androidTest/kotlin/io/realm/PrimaryKeyTests.kt
