+++ test/src/macosTest/kotlin/io/realm/shared/ImportTests.kt
@@ -1,0 +1,1 @@
+test/src/androidTest/kotlin/io/realm/shared/ImportTests.kt
